+++ db-tool/src/app/app.module.ts
@@ -6,7 +6,7 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 
 import { RoomSelectionComponent } from './room-selection.component';
 import { ConfigurationComponent } from './configuration.component';
-import { RoomEditorComponent } from './devices.component';
+import { RoomEditorComponent } from './room-editor.component';
 import { AddBuildingComponent } from './add-building.component';
 import { AddRoomComponent } from './add-room.component';
