+++ db-tool/src/app/room-editor.component.html
